Move printe.py to import_bots and update imports
Relocated printe.py from utils to import_bots and updated all relevant import statements to use the new location. Removed the now-unused utils/__init__.py file.
