+++ ncc_core/nc_import/import_bots/import_files.py
@@ -4,7 +4,7 @@ bot for importing files from nccommons to wikipedia
 
 """
 import re
-from ..utils import printe
+from . import printe
 from . import upload_file
 from .db import add_to_db, add_to_jsonl
 from .ncc_page import load_main_api
