--- ncc_core/nc_import/utils/__init__.py
@@ -1,1 +1,0 @@
-
